WAT-1169 move assert into own package {async-assert}
