+++ packages/async-assert/.npmrc
@@ -1,0 +1,3 @@
+package-lock=false
+save-exact=true
+
