+++ packages/async-assert/README.md
@@ -1,0 +1,11 @@
+# `assert`
+
+> TODO: description
+
+## Usage
+
+```
+const assert = require('assert');
+
+// TODO: DEMONSTRATE API
+```
