+++ packages/async-assert/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "async-assert",
+  "version": "0.5.35",
+  "description": "an async wrapper around chai",
+  "keywords": [
+    "assert"
+  ],
+  "author": "Sergei.Zhurin <Sergei.Zhurin@nordigy.ru>",
+  "homepage": "https://github.com/ringcentral/testring#readme",
+  "license": "ISC",
+  "main": "./dist/index.js",
+  "types": "./src/index.ts",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ringcentral/testring.git"
+  },
+  "scripts": {
+    "test": "echo \"Error: run tests from root\" && exit 1"
+  },
+  "bugs": {
+    "url": "https://github.com/ringcentral/testring/issues"
+  },
+  "dependencies": {
+    "@testring/types": "^0.5.35",
+    "chai": "^4.3.3"
+  }
+}
