+++ packages/async-assert/tsconfig.json
@@ -1,0 +1,38 @@
+{
+  "compilerOptions": {
+    "target": "es2015",
+    "module": "commonjs",
+    "allowJs": false,
+    "moduleResolution": "node",
+    "noFallthroughCasesInSwitch": true,
+    "allowSyntheticDefaultImports": true,
+    "preserveConstEnums": true,
+    "noEmitOnError": true,
+    "declaration": false,
+    "outDir": "dist",
+    "sourceMap": false,
+    "lib": [
+      "es5",
+      "es2015",
+      "es2016",
+      "es2017",
+      "dom"
+    ],
+    "typeRoots": [
+      "./node_modules/@types",
+      "../../node_modules/@types"
+    ],
+    "noUnusedLocals": true,
+    "strictNullChecks": true,
+    "removeComments": false
+  },
+  "include": [
+    "*.ts",
+    "src/**/*.ts",
+    "src/**/*.json"
+  ],
+  "exclude": [
+    "./node_modules",
+    "../../node_modules"
+  ]
+}
