+++ packages/types/src/index.ts
@@ -1,3 +1,5 @@
+export * from './async-assert';
+
 export * from './cli';
 
 export * from './test-run-controller';
@@ -45,4 +47,5 @@ export * from './dependencies-builder';
 export * from './utils';
 
 export * from './web-application/enums';
+
 export * from './web-application';
