+++ packages/types/src/web-application/index.ts
@@ -23,25 +23,6 @@ export interface IWebApplicationResponseMessage {
     error: Error | null;
 }
 
-export interface IAssertionSuccessMeta {
-    isSoft: boolean;
-    successMessage?: string;
-    assertMessage?: string;
-    originalMethod: string;
-    args: any[];
-}
-
-export interface IAssertionErrorMeta extends IAssertionSuccessMeta {
-    errorMessage?: string;
-    error?: Error;
-}
-
-export interface IAssertionOptions {
-    isSoft?: boolean;
-    onSuccess?: (IAssertionSuccessMeta) => void | Promise<void>;
-    onError?: (IAssertionErrorMeta) => void | Error | Promise<void | Error>;
-}
-
 export type IWebApplicationClient = {
     [K in keyof IBrowserProxyPlugin]: (...args: Array<any>) => Promise<any>
 };
