+++ packages/web-application/src/index.ts
@@ -1,3 +1,2 @@
-export { createAssertion } from './assert';
 export { WebApplicationController } from './web-application-controller';
 export { WebApplication } from './web-application';
